requirements.txt
Flask>=2.2.5
Flask-SocketIO>=5.1.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python>=8.0.26
cryptography>=3.4.7
python-dotenv>=0.19.0
bcrypt>=4.1.2
scapy>=2.4.5
